chore: Resolve compilation errors
