+++ crates/hyperswitch_interfaces/Cargo.toml
@@ -26,6 +26,7 @@ thiserror = "1.0.58"
 time = "0.3.35"
 
 # First party crates
+actix-web = "4.5.1"
 api_models = { version = "0.1.0", path = "../api_models" }
 common_enums = { version = "0.1.0", path = "../common_enums" }
 common_utils = { version = "0.1.0", path = "../common_utils" }
