+++ crates/router/src/types/api/webhooks.rs
@@ -1,1 +1,5 @@
-pub use hyperswitch_interfaces::webhooks::IncomingWebhook;
+pub use api_models::webhooks::{
+    AuthenticationIdType, IncomingWebhookDetails, IncomingWebhookEvent, MerchantWebhookConfig,
+    ObjectReferenceId, OutgoingWebhook, OutgoingWebhookContent, WebhookFlow,
+};
+pub use hyperswitch_interfaces::webhooks::{IncomingWebhook, IncomingWebhookRequestDetails};
